services/vanna/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
vanna==0.5.5
groq==0.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0